+++ packages/ondes_sdk/lib/ondes_sdk.dart
@@ -33,6 +33,7 @@ export 'src/modules/storage.dart';
 export 'src/modules/app.dart';
 export 'src/modules/friends.dart';
 export 'src/modules/social.dart';
+export 'src/modules/websocket.dart';
 
 // Models
 export 'src/models/user_profile.dart';
@@ -46,6 +47,7 @@ export 'src/models/comment.dart';
 export 'src/models/story.dart';
 export 'src/models/media.dart';
 export 'src/models/social_user.dart';
+export 'src/models/websocket_connection.dart';
 
 // Exceptions
 export 'src/models/exceptions.dart';